pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rocky-agent"
version = "0.1.0"
description = "Rocky: CLI-first, file-first, continuously learning general agent"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "OpenAI" }]
dependencies = [
  "httpx>=0.27",
  "prompt_toolkit>=3.0.50",
  "rich>=14.1",
  "PyYAML>=6.0",
  "beautifulsoup4>=4.12",
  "openpyxl>=3.1",
  "pandas>=2.2",
  "playwright>=1.50",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[project.scripts]
rocky = "rocky.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rocky = ["data/bundled_skills/**/SKILL.md"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
